felsch-tree: separate out from ToddCoxeter
