+++ Makefile.am
@@ -66,6 +66,7 @@ pkginclude_HEADERS += include/libsemigroups/digraph-with-sources.tpp
 pkginclude_HEADERS += include/libsemigroups/digraph.hpp
 pkginclude_HEADERS += include/libsemigroups/exception.hpp
 pkginclude_HEADERS += include/libsemigroups/fastest-bmat.hpp
+pkginclude_HEADERS += include/libsemigroups/felsch-tree.hpp
 pkginclude_HEADERS += include/libsemigroups/forest.hpp
 pkginclude_HEADERS += include/libsemigroups/fpsemi-intf.hpp
 pkginclude_HEADERS += include/libsemigroups/fpsemi.hpp
@@ -552,6 +553,7 @@ libsemigroups_la_SOURCES += src/cong-pair.cpp
 libsemigroups_la_SOURCES += src/cong.cpp
 libsemigroups_la_SOURCES += src/constants.cpp
 libsemigroups_la_SOURCES += src/coset.cpp
+libsemigroups_la_SOURCES += src/felsch-tree.cpp
 libsemigroups_la_SOURCES += src/forest.cpp
 libsemigroups_la_SOURCES += src/fpsemi-intf.cpp
 libsemigroups_la_SOURCES += src/fpsemi.cpp
